code/requirements.txt
google-genai>=1.16.0
python-dotenv>=1.0.0
pandas>=2.0.0
rank-bm25>=0.2.2
langdetect>=1.0.9
